requirements.txt
jsonschema==4.23.0
openapi-spec-validator==0.7.1
PyYAML==6.0.2